Added private variables for myHand and myBoard for the AI.java so that it could pass it to AIMove. These variables are private and have setters for them. Added more into move function in AI.java, including calls to accessAllCards, setMyHand, setBoard, creating an instance of AIMove and feeding it the info it needs, and initializing a string that will hold the card the AIMove will choose.
Added private variables allCards, board, numPlayers into AIMove so that it can determine the best moves given all that information.
Updated the constructor to use the setters for the private variables in AIMove.

Started chooseCard method which will use the information passed into AIMove to return a String containing the best card to choose for that turn.
Created setters for the private variables in AIMove

Slowly starting to remove functions from HardComputer.